pyproject.toml
[project]
name = "mr-index-linking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "pymupdf>=1.26.7",
    "rich>=14.2.0",
]
